refactor(terminal): extract TerminalRevealController
- Move setVisible, wake, wakeForFocus, fullWakeForVisibilityRestore,
  repaintForReveal, reconcileRevealGeometry, and revealTerminal into a new
  sibling controller, line-for-line — setVisible straddles WebGL + resize
  + reflow, and two past incidents (#5878 want-membership, #7762
  backgrounded-terminal geometry corruption) shipped in exactly this
  ordering, so call order is preserved verbatim
- hostHasRenderableDims, ensureOpened, handlePostWake,
  deferGridChangeForStream, and checkStaleDirecting stay on the service
  (used elsewhere too) and are passed in as Deps callbacks
- All seven methods stay public delegates on TerminalInstanceService —
  reached externally via UI components, store slices, and
  PaintFabricCompositor's structural TerminalPaintPlane contract
- Fix a lockResize deps-forwarding bug that always passed an explicit
  third argument even when the caller omitted it
- Update tests that spied on the service's own methods to intercept the
  controller's internal self-calls (wake -> fullWake/repaintForReveal,
  revealTerminal -> fullWake/repaintForReveal) to spy on revealController
  instead, since that's where the call graph now lives
